+++ js/util.js
@@ -1,0 +1,18 @@
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+const getElement = (array) => array[getRandomInteger(0, array.length - 1)];
+
+export {createIdGenerator, getRandomInteger, getElement};
